requirements.txt
wheel==0.45.1
langchain==0.2.14
langgraph==0.2.16
langchain-google-genai==1.0.7

fastapi==0.112.0
uvicorn==0.30.6

chromadb==0.5.5

python-dotenv==1.0.1

pydantic==2.8.2